+++ InputGlyphs/Assets/InputGlyphs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.6.0] - 2024-07-05
+### Changed
+- Change GetGlyph() to LoadGlyph().
+  - Texture2D GetGlyph(devices, path) -> bool LoadGlyph(texture, devices, path)
+  - Fixed memory allocations and leaks caused by regenerating textures.
+
 ## [0.5.2] - 2024-07-05
 ### Fixed
 - Fix detection player input changes in TextGlyph